Implement abstract voter and remove operations
